fixed saving dosages
Not really fixed because it never actually worked but basically I just created a dictionary for the medicine and dosages taken for the day and then saved it into dosages_track.txt. Whenever the website is opened again, it would read dosages_track.txt and convert it into a dictionary.